requirements.txt
pytz==2023.3
requests==2.30.0
launchd==0.3.0; sys_platform == 'darwin'
python-dotenv==1.0.0
